[E2E] Test with plugin text on agreements modal form is added
